Extract DAQI properties to a module
And switch refactor PollenPrediction to use it.
